Adding some test-paths to deploy script to prevent unnecessary errors.

+++ Matt-Kwiecien-MVC/Deploy/deploy.ps1
@@ -2,10 +2,14 @@
 Compress-Archive -Path ((Get-Item -Path ".\").FullName + "\..\bin\debug\netcoreapp2.1\") -DestinationPath site.zip
 
 #Remove old deploy package
-rm matt-kwiecien-mvc.zip
+if (Test-Path matt-kwiecien-mvc.zip) {
+    rm matt-kwiecien-mvc.zip
+}
 
 #Create new deploy package
 Compress-Archive -Path site.zip, aws-windows-deployment-manifest.json -DestinationPath matt-kwiecien-mvc.zip
 
 #Remove site zip
-rm site.zip
+if (Test-Path site.zip) {
+    rm site.zip
+}
